+++ tui/enet/net.cpp
@@ -277,21 +277,21 @@ void deregister_object(InventoryElement * o)
     objects.remove(obj);
 }
 
-BaseElement * get_base_element(int32_t id)
+BaseElement * get_base_element(size_t id)
 {
     BaseListElement *el=(BaseListElement*)base_elements.find(&id);
     if (!el) return nullptr;
     return (BaseElement*)((el)->base);
 }
 
-BasePlant * get_base_plant(int32_t id)
+BasePlant * get_base_plant(size_t id)
 {
     BaseListElement *el=(BaseListElement*)base_plants.find(&id);
     if (!el) return nullptr;
     return (BasePlant*)((el)->base);
 }
 
-BaseAnimal * get_base_animal(int32_t id)
+BaseAnimal * get_base_animal(size_t id)
 {
     BaseListElement *el=(BaseListElement*)base_animals.find(&id);
     if (!el) return nullptr;
